Rewrite grabbing: velocity for movement, E to grab
solves #3

All grabbables now need a GrabbedController script attached to them!

One of the boxes in the Copy level (the box lying on ground) has this
already.

Grabbing now uses E, same as Interact, through the IInteractable iface.

It's kinda janky that I did all this in one single file without changing
any of our other scripts or any of our input logic. The
PlayerInteract.cs script is really not made for interacting with objects
where the interaction happens over time with several instances of input
(i.e. grab and then ungrab on the same object, without having to look at
it to ungrab)

We can remove the old GrabManager.cs when we merge this. Except:

One thing missing compared to GrabManager is that objects are not
automatically dropped if they go out of range or start to be rewinded!

Grabbed objects are now moved by setting a constant velocity instead of
applying force. This makes it a lot snappier than before.

One thing that is still a bit shite is that grabbed objects slightly
jitter (much less than before) while walking with them in your hand. I
think this is due to their physics updating at a slightly different
timing to the player. I tried various ways of setting the object
position and all had this problem: rb.position, rb.MovePosition(), and
lastly rb.velocity. I think the only way to fix this might be to either
go deep in the physics simulation logic and fix the timings, or we make
grabbed objects a child of the player (but then how handle collisions?).
I also tried to mess with interpolation settings on grabbed objects to
see if it solved the jittering, which it did not seem to do.

To fix issues with friction (objects dragged along floor) when moving
while holding I now clamp grabbed objects position to slightly above the
player's feet.

And also grabbing of objects now takes into account their size, so you
maintain a distance of 2.0f to the nearest edge of the grabbed object
instead of the center! This lets you grab big stuff without impeding
yourself.

The grab-flying exploit is still in the game, but much harder to do. The
only reason it still works I believe is because you can jump on an
object while holding it (Newton's whichever law it was hello????). So we
just gotta check that in the jump function ig.

Oh and excuse the variable and function names.