you can escape!
Though a couple things:

No sounds on button or hatch

Idk how to make a mesh for the button, the backplate is what you're
technically interacting with (but it works really well like that tbh)

The hatch animation has some weird skips (I think we had that happen for
the door in our first demo some times?) idk why tho

Rewinding the hatch is a bit strange, hovering over any part of the
hatch should focus the entire hatch since that is what has the
TimeEntity component. So we should probably change the logic for
selecting objects to rewind.

The animation is a bit quick so it might be a bit too difficult right
now?

The button is not animated yet (just make an animation and trigger it on
click though, not the most complex thing to do)